src/requirements.txt
hidapi
pystray
keyboard
customtkinter
sounddevice
pillow
pyinstaller